# docs/mobile-api.openapi.yaml
openapi: 3.0.3
info:
  title: SchoolEdge Mobile API
  version: "1.0"
  description: |
    Stateless bearer-token REST API backing the SchoolEdge Flutter app (both
    the shared multi-tenant "SaaS" app and per-school branded builds), bolted
    onto the existing CodeIgniter 3 session-based web app. Every endpoint here
    lives under `application/controllers/api/v1/` and is routed in
    `application/config/routes.php`; all of them extend `Api_Controller`
    (`application/core/Api_Controller.php`), which is the source of truth for
    the auth/error/envelope conventions documented below.

    **Multi-tenancy**: a `branch` row is the sole tenant unit. Every
    authenticated request resolves to one `mobile_memberships` row (a
    login_credential + branch + role triple), and every endpoint scopes its
    queries to that membership's `branch_id` - there is no cross-branch access
    by design (see `TenantIsolationTest` in `tests/Api/`).

    **Read-only demo branches**: a branch with `is_demo=1` allows every GET but
    rejects every mutating request with `demo_readonly` (403) -
    `Api_Controller::blockIfDemoReadonly()`.
  contact:
    name: SchoolEdge
servers:
  - url: https://{host}/api/v1/mobile
    variables:
      host:
        default: schooledge.ng
        description: Replace with the target install's own host (each school's own domain for a branded app, or the shared SaaS host).
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT-like
      description: >
        A short-lived (15 min) HMAC-signed access token from `auth/login` /
        `auth/refresh` - not a standard JWT, but the same base64-payload +
        signature shape. Every endpoint except `auth/*` and `schools` requires
        `Authorization: Bearer <access_token>`.
  schemas:
    Envelope:
      type: object
      description: Every response (success or failure) uses this shape - see `Api_Controller::ok()`/`fail()`.
      properties:
        success:
          type: boolean
        data:
          nullable: true
        meta:
          type: object
        error:
          type: object
          nullable: true
          properties:
            code:
              type: string
              example: validation_error
            message:
              type: string
            details:
              nullable: true
    Error401:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      example:
        success: false
        error: { code: unauthorized, message: "A bearer token is required.", details: null }
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with username/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
                membership_id: { type: integer, description: pick a specific membership when the credential has more than one active branch/role }
                installation_id: { type: string, description: "stable per-device id generated by the client, persisted across logout" }
                platform: { type: string, example: android }
                app_version: { type: string }
      responses:
        '200':
          description: Logged in directly, or a 2FA challenge is required
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                example:
                  success: true
                  data:
                    tokens: { access_token: "...", refresh_token: "...", token_type: Bearer, expires_in: 900 }
                    membership: { id: 1, status: active, is_default: true, school: {}, role: {} }
                    memberships: []
        '401':
          description: Invalid credentials
          content: { application/json: { schema: { $ref: '#/components/schemas/Error401' } } }
        '403':
          description: No active membership for this credential
        '429':
          description: Rate limited (10 login attempts/minute/IP)
...